Made it such that the nodes follow the transformComponent, fixed warnings, added invalid entity in pch.h

+++ BoofWoof/Core/EngineCore.cpp
@@ -257,7 +257,7 @@ void EngineCore::OnUpdate()
 
 	{
 		// Pathfinding
-		mPathfindingSys->Update(m_DeltaTime);
+		mPathfindingSys->Update(static_cast<float>(m_DeltaTime));
 	}
 
 	static bool drawTestLinesOnce = true;
